+++ CHANGES.rst
@@ -15,8 +15,12 @@ Imviz
 Mosviz
 ^^^^^^
 
+- New metadata viewer plugin. [#1035]
+
 Specviz
 ^^^^^^^
+
+- New metadata viewer plugin. [#1035]
 
 API Changes
 -----------